Refactor file imports and add file exporter functionality for IAGA data handling
