+++ backend/src/usecase/ee_index/calc_moving_avg.py
