+++ backend/src/usecase/file_exporter/remove_files.py
@@ -8,5 +8,6 @@ TMP_DIR_PATH = "/tmp"
 
 def remove_tmp_files():
     tmp_dir_path = generate_parent_abs_path(TMP_DIR_PATH)
-    shutil.rmtree(tmp_dir_path)
+    if os.path.exists(tmp_dir_path):
+        shutil.rmtree(tmp_dir_path)
     os.mkdir(tmp_dir_path)
